requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Jinja2==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
